-I option in Make file, optimization of unsigned long long

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: lannur-s <lannur-s@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/13 18:04:15 by lannur-s          #+#    #+#             */
-/*   Updated: 2023/06/21 12:05:42 by lannur-s         ###   ########.fr       */
+/*   Updated: 2023/08/01 15:55:51 by lannur-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,8 @@
 # define FT_PRINTF_H
 
 # include <stdarg.h>
-# include <stddef.h> 
-# include <stdio.h>
 # include <limits.h>
-# include "./libft/libft.h"
+# include "libft.h"
 
 # define STD_OUT 1
 # define BASE_HEX 16
@@ -31,6 +29,6 @@ int		ft_printf(const char *format, ...);
 int		print_char(char c);
 int		print_string(char *s);
 int		print_nbr_signed(int n);
-int		print_nbr_base(unsigned long long n, int base, int uppercase);
+int		print_nbr_base(unsigned long n, int base, int uppercase);
 
 #endif
