+++ print_nbr_base.c
@@ -6,13 +6,13 @@
 /*   By: lannur-s <lannur-s@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/15 14:34:57 by lannur-s          #+#    #+#             */
-/*   Updated: 2023/06/21 13:07:43 by lannur-s         ###   ########.fr       */
+/*   Updated: 2023/08/01 15:49:44 by lannur-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	print_nbr_base(unsigned long long nbr, int base, int uppercase)
+int	print_nbr_base(unsigned long nbr, int base, int uppercase)
 {
 	int				length;
